Close the remaining #1447 gaps, harden the low preset, fix main's red lanes
Architecture (issue #1447):
- Render: seed Carthaginian elephants from the EntityID, not the Entity*
  address; one GL lifetime policy (textures, shaders, the wear volume and
  meshes release or defer by share group, never call GL without a context);
  meshes re-upload in a foreign share group; atomic GraphicsSettings;
  shader reload skips already-compiled and foreign-group programs; the
  runtime-bake barrier is lifted only by the renderer that raised it; the
  portrait renderer gets its own frame profile; a debug freeze asserts no
  component layout change during parallel prepare.
- Sim: replay digest gains a "systems" part built from opt-in SessionSnapshot
  digest views (AI, undead zones, veins, ground wildlife); replay format 5.
  Victory, mission-wave and wildlife clocks accumulate in double (a float
  running sum drifted ~0.16 s per 1000 s). MapTransformer resolves owners,
  nations, terrain, collision and factories through the world's bound
  services; session_for counts unbound fallbacks. Dropped ticks show on F10.
- Determinism: TerrainService revisions are process-unique, so a second match
  can no longer reuse the previous map's nav grid / prop index (this was why
  AiDuelMatchTest.HannibalMirror failed after another AI match in the same
  binary); the RPG crit roll and FPV swing variant no longer use process-wide
  sequences; TroopProfileService is thread-safe (render thread raced the sim).

Performance (digests bit-identical before/after on 13 scenarios):
- exact squared-distance prefilters in walk_formation_slot and body impact;
  per-tick formation rank roster instead of O(F^2) scans; layout copies
  reuse capacity; troop profiles by reference; gate loop hoisting; cull
  radius memoised; horse mount frame samples only the two joints it needs.
  sim_benchmark 2000 units: ~41 -> ~37 ms/tick interleaved A/B.

Low preset / old GPUs:
- first run picks Low on software renderers and sub-4.3 GL (not on macOS);
  Low uses an R11F_G11F_B10F post target, 4 local lights, smaller terrain
  atlases; bigger palette ring for the per-creature fallback; unsynchronised
  fenced fallback mapping; validator check that 4.x calls stay gated.
  Verified: renders and presents on Mesa llvmpipe capped at GL 3.3 Core.

Fixed on main: RpgFpvOverlay imported its own module (arena HUD failed to
load); 20 uncatalogued overlay strings translated in all eight languages;
UBSan downcast in gl_lifetime; PR lane now runs a short headless replay.

+++ game/mission/mission_wave_runtime.cpp
@@ -19,7 +19,7 @@ auto MissionWaveRuntime::owner_has_unspawned_waves(int owner_id) const -> bool {
 }
 
 void MissionWaveRuntime::reset() {
-  m_elapsed = 0.0F;
+  m_elapsed = 0.0;
   m_waves.clear();
   m_events.clear();
   m_director.reset();
@@ -32,7 +32,7 @@ void MissionWaveRuntime::bind_after_setup(
   m_waves = std::move(waves);
   m_events = std::move(events);
   m_director.bind(&m_waves, binding.world);
-  m_director.set_elapsed(m_elapsed);
+  m_director.set_elapsed(static_cast<float>(m_elapsed));
   if (binding.victory_service != nullptr) {
     binding.victory_service->set_mission_wave_query(&m_director);
   }
@@ -68,7 +68,7 @@ void MissionWaveRuntime::restore(const MissionWaveBinding& binding,
   m_director.restore(wave_state);
   m_elapsed = m_director.elapsed();
   for (auto& event : m_events) {
-    event.fired = m_elapsed >= event.trigger_time;
+    event.fired = m_elapsed >= static_cast<double>(event.trigger_time);
   }
 
   if (binding.victory_service != nullptr) {
@@ -79,7 +79,7 @@ void MissionWaveRuntime::restore(const MissionWaveBinding& binding,
 auto MissionWaveRuntime::fire_due_events() -> QStringList {
   QStringList announcements;
   for (auto& event : m_events) {
-    if (event.fired || m_elapsed < event.trigger_time) {
+    if (event.fired || m_elapsed < static_cast<double>(event.trigger_time)) {
       continue;
     }
     event.fired = true;
@@ -98,20 +98,21 @@ auto MissionWaveRuntime::advance(const MissionWaveBinding& binding,
     return effects;
   }
 
-  m_elapsed += delta_seconds;
+  m_elapsed += static_cast<double>(delta_seconds);
   effects.announcements = fire_due_events();
 
   if (m_waves.empty()) {
     return effects;
   }
 
-  m_director.set_elapsed(m_elapsed);
+  m_director.set_elapsed(static_cast<float>(m_elapsed));
   const auto director_effects = m_director.advance();
 
   bool spawned_any = false;
   for (const auto index : director_effects.waves_to_spawn) {
     const auto spawn_effects =
-        m_spawner.spawn({*binding.world, *binding.level, m_elapsed}, m_waves[index]);
+        m_spawner.spawn({*binding.world, *binding.level, static_cast<float>(m_elapsed)},
+                        m_waves[index]);
     for (const auto& announcement : spawn_effects.mission_announcements) {
       effects.announcements.append(announcement);
     }
